piper_requirements.txt
flask==3.0.0
piper-tts==1.2.0